+++ src/pages/index.tsx
@@ -1,10 +1,10 @@
 import * as React from "react"
-import { Container } from "@mui/material"
+import Container from "@mui/material/Container"
 import Layout from "../components/Layout"
-import About from "../components/About"
-import Experience from "../components/Experience"
-import Projects from "../components/Projects"
-import Contact from "../components/Contact"
+import About from "../components/sections/About"
+import Experience from "../components/sections/Experience"
+import Projects from "../components/sections/Projects"
+import Contact from "../components/sections/Contact"
 
 const IndexPage = () => {
   return (
